+++ src/version2/issueTypeScreenSchemes.ts
@@ -44,6 +44,9 @@ export class IssueTypeScreenSchemes {
         startAt: parameters?.startAt,
         maxResults: parameters?.maxResults,
         id: parameters?.id,
+        queryString: parameters?.queryString,
+        orderBy: parameters?.orderBy,
+        expand: parameters?.expand,
       },
     };