+++ src/version2/models/group.ts
@@ -3,6 +3,11 @@ import { PagedListUserDetailsApplicationUser } from './pagedListUserDetailsAppli
 export interface Group {
   /** The name of group. */
   name?: string;
+  /**
+   * The ID of the group, which uniquely identifies the group across all Atlassian products. For example,
+   * *952d12c3-5b5b-4d04-bb32-44d383afc4b2*.
+   */
+  groupId?: string;
   /** The URL for these group details. */
   self?: string;
   users?: PagedListUserDetailsApplicationUser;