+++ src/version2/models/paginatedResponseComment.ts
@@ -2,7 +2,7 @@ import { Comment } from './comment';
 
 export interface PaginatedResponseComment {
   total?: number;
-  results?: Comment[];
-  maxResults?: number;
   startAt?: number;
+  maxResults?: number;
+  results?: Comment[];
 }