+++ src/version2/parameters/addGadget.ts
@@ -1,0 +1,6 @@
+import { DashboardGadgetSettings } from '../models';
+
+export interface AddGadget extends DashboardGadgetSettings {
+  /** The ID of the dashboard. */
+  dashboardId: number;
+}