+++ src/version2/parameters/getProjectComponentsPaginated.ts
@@ -1,20 +1,35 @@
 export interface GetProjectComponentsPaginated {
-  /** The project ID or project key (case sensitive). */
+  /** The project ID or project key (case-sensitive). */
   projectIdOrKey: string;
   /** The index of the first item to return in a page of results (page offset). */
   startAt?: number;
   /** The maximum number of items to return per page. */
   maxResults?: number;
   /**
-   * [Order](#ordering) the results by a field:
+   * [Order](https://developer.atlassian.com/cloud/jira/platform/rest/v2/intro/#ordering) the results by a field:
    *
-   * `description` Sorts by the component description. `issueCount` Sorts by the count of issues associated with the
-   * component. `lead` Sorts by the user key of the component's project lead. `name` Sorts by component name.
+   * - `description` Sorts by the component description.
+   * - `issueCount` Sorts by the count of issues associated with the component.
+   * - `lead` Sorts by the user key of the component's project lead.
+   * - `name` Sorts by component name.
    */
-  orderBy?: string;
+  orderBy?:
+  | 'description'
+  | 'issueCount'
+  | 'lead'
+  | 'name'
+  | '+description'
+  | '+issueCount'
+  | '+lead'
+  | '+name'
+  | '-description'
+  | '-issueCount'
+  | '-lead'
+  | '-name'
+  | string;
   /**
-   * Filter the results using a literal string. Components with a matching `name` or `description` are returned (case
-   * insensitive).
+   * Filter the results using a literal string. Components with a matching `name` or `description` are returned
+   * (case-insensitive).
    */
   query?: string;
 }