+++ src/version2/parameters/getWorkflowsPaginated.ts
@@ -23,7 +23,25 @@ export interface GetWorkflowsPaginated {
    * - `default` For each workflow, returns information about whether this is the default workflow.
    * - `schemes` For each workflow, returns information about the workflow schemes the workflow is assigned to.
    */
-  expand?: string;
+  expand?:
+  | 'transitions'
+  | 'transitions.rules'
+  | 'transitions.properties'
+  | 'statuses'
+  | 'statuses.properties'
+  | 'default'
+  | 'schemes'
+  | (
+    | 'transitions'
+    | 'transitions.rules'
+    | 'transitions.properties'
+    | 'statuses'
+    | 'statuses.properties'
+    | 'default'
+    | 'schemes'
+  )[]
+  | string
+  | string[];
   /** String used to perform a case-insensitive partial match with workflow name. */
   queryString?: string;
   /**
@@ -33,7 +51,17 @@ export interface GetWorkflowsPaginated {
    * - `created` Sorts by create time.
    * - `updated` Sorts by update time.
    */
-  orderBy?: 'name' | 'created' | 'updated' | string;
+  orderBy?:
+  | 'name'
+  | 'created'
+  | 'updated'
+  | '+name'
+  | '+created'
+  | '+updated'
+  | '-name'
+  | '-created'
+  | '-updated'
+  | string;
   /** Filters active and inactive workflows. */
   isActive?: boolean;
 }