+++ src/version2/parameters/removeUserFromGroup.ts
@@ -1,6 +1,11 @@
 export interface RemoveUserFromGroup {
-  /** The name of the group. */
-  groupname: string;
+  /**
+   * As a group's name can change, use of `groupId` is recommended to identify a group. The name of the group. This
+   * parameter cannot be used with the `groupId` parameter.
+   */
+  groupname?: string;
+  /** The ID of the group. This parameter cannot be used with the `groupName` parameter. */
+  groupId?: string;
   /**
    * This parameter is no longer available. See the [deprecation
    * notice](https://developer.atlassian.com/cloud/jira/platform/deprecation-notice-user-privacy-api-migration-guide/)