+++ src/version2/parameters/updateGadget.ts
@@ -1,0 +1,8 @@
+import { DashboardGadgetUpdateRequest } from '../models';
+
+export interface UpdateGadget extends DashboardGadgetUpdateRequest {
+  /** The ID of the dashboard. */
+  dashboardId: number;
+  /** The ID of the gadget. */
+  gadgetId: number;
+}